+++ apps/docs/app/lib/formatingCode.ts
@@ -1,0 +1,9 @@
+import { highlightCode } from "@/components/highlightCode";
+
+export async function formatCode(code: string, language: string) {
+    return await highlightCode(`
+\`\`\`${language}
+${code}
+\`\`\`
+`);
+}