+++ packages/components/src/utils/src/types.ts
@@ -20,19 +20,21 @@ export interface DOMProps extends StyleProps, SharedDOMProps {
 
 export interface StyleRenderProps<T> {
     /** The CSS [className](https://developer.mozilla.org/en-US/docs/Web/API/Element/className) for the element. A function may be provided to compute the class based on component state. */
-    className?: string | ((values: T) => string);
+    className?: string | ((values: T & { defaultClassName: string | undefined }) => string);
     /** The inline [style](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/style) for the element. A function may be provided to compute the style based on component state. */
-    style?: CSSProperties | ((values: T) => CSSProperties);
+    style?: CSSProperties | ((values: T & { defaultStyle: CSSProperties }) => CSSProperties);
 }
+
 export interface RenderProps<T> extends StyleRenderProps<T> {
     /** The children of the component. A function may be provided to alter the children based on component state. */
-    children?: ReactNode | ((values: T) => ReactNode);
+    children?: ReactNode | ((values: T & { defaultChildren: ReactNode | undefined }) => ReactNode);
 }
 
 export interface RenderPropsHookOptions<T> extends RenderProps<T>, SharedDOMProps, AriaLabelingProps {
     values: T;
     defaultChildren?: ReactNode;
     defaultClassName?: string;
+    defaultStyle?: CSSProperties;
 }
 
 // TODO: i added this, not sure if it will make sense