+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,161 @@
 # @hopper-ui/icons
 
+## 2.1.1
+
+### Patch Changes
+
+- a92e653: fix inversion between background and shadow values
+
+## 2.1.0
+
+### Minor Changes
+
+- 75db68e: Added the following SVG icons:
+
+  **Added**
+
+  | SVG Name              |
+  | --------------------- |
+  | graduation-hat-16.svg |
+  | graduation-hat-24.svg |
+  | graduation-hat-32.svg |
+  | image-16.svg          |
+  | image-24.svg          |
+  | image-32.svg          |
+  | screen-share-16.svg   |
+  | screen-share-24.svg   |
+  | screen-share-32.svg   |
+
+  Added the following Icons:
+
+  **Added**
+
+  | Icon Name       |
+  | --------------- |
+  | GraduationHat   |
+  | ImageIcon       |
+  | ScreenShareIcon |
+
+## 2.0.0
+
+### Major Changes
+
+- c5c6b30: ## @hopper-ui/tokens
+
+  **Added**
+
+  | Token                                 |
+  | ------------------------------------- |
+  | danger-text-selected                  |
+  | danger-border-selected                |
+  | danger-icon-selected                  |
+  | danger-icon-weak-hover                |
+  | danger-icon-weak-press                |
+  | danger-surface-selected               |
+  | danger-surface-weak-hover             |
+  | danger-surface-weak-press             |
+  | danger-text-weak-hover                |
+  | danger-text-weak-press                |
+  | decorative-option5-surface-weak-hover |
+  | neutral-surface-weak-selected         |
+  | neutral-surface-weakest-selected      |
+  | neutral-border-selected               |
+  | neutral-icon-selected                 |
+  | neutral-icon-strong-hover             |
+  | neutral-icon-weak-press               |
+  | neutral-surface-selected              |
+  | neutral-text-weak-hover               |
+  | neutral-text-weak-press               |
+  | primary-border-selected               |
+  | primary-icon-selected                 |
+  | primary-surface-selected              |
+  | primary-surface-strong-selected       |
+  | primary-surface-weak-hover            |
+  | primary-surface-weak-press            |
+  | primary-text-selected                 |
+  | upsell-border-selected                |
+  | upsell-icon-selected                  |
+  | upsell-icon-hover                     |
+  | upsell-icon-weak-hover                |
+  | upsell-icon-weak-press                |
+  | upsell-surface-selected               |
+  | upsell-surface-weak-hover             |
+  | upsell-surface-weak-press             |
+  | upsell-text-selected                  |
+  | upsell-text-weak                      |
+  | upsell-text-weak-hover                |
+  | upsell-text-weak-press                |
+
+  **Renamed**
+
+  Use this as a reference to fix all breaking changes.
+
+  | Old Token name                | New Token name               |
+  | ----------------------------- | ---------------------------- |
+  | danger-border-active          | danger-border-press          |
+  | danger-icon-active            | danger-icon-press            |
+  | danger-text-active            | danger-text-press            |
+  | danger-surface-active         | danger-surface-press         |
+  | neutral-border-active         | neutral-border-press         |
+  | neutral-icon-active           | neutral-icon-press           |
+  | neutral-surface-active        | neutral-surface-press        |
+  | neutral-surface-weak-active   | neutral-surface-weak-press   |
+  | neutral-text-active           | neutral-text-press           |
+  | primary-border-active         | primary-border-press         |
+  | primary-icon-active           | primary-icon-press           |
+  | primary-surface-active        | primary-surface-press        |
+  | primary-text-active           | primary-text-press           |
+  | primary-surface-strong-active | primary-surface-strong-press |
+  | upsell-border-active          | upsell-border-press          |
+  | upsell-icon-active            | upsell-icon-press            |
+  | upsell-surface-active         | upsell-surface-press         |
+  | upsell-text-active            | upsell-text-press            |
+
+  **Updated**
+
+  | Token                        |
+  | ---------------------------- |
+  | danger-surface-hover         |
+  | danger-surface-strong        |
+  | danger-surface-strong-hover  |
+  | danger-text-weak             |
+  | status-progress-text         |
+  | status-progress-icon         |
+  | primary-surface-strong-hover |
+  | primary-text                 |
+  | primary-surface-hover        |
+  | upsell-surface-hover         |
+  | upsell-surface-weak          |
+  | upsell-text                  |
+  | upsell-icon                  |
+  | upsell-text-hover            |
+
+  ## @hopper-ui/components
+
+  - Added a press state to Links
+  - Added a press state to Buttons
+  - Added a press state to Checkbbox
+  - Added a press state to Switch
+  - Added a press state to Radio
+  - Added numerous tests to test Pressed States of components
+
+### Patch Changes
+
+- 73037f3: Updated the icons variables to be more consistent with the other components.
+
+  Will need to update the style sheets to use the new CSS variables.
+
+  For example, they will need to replace:
+
+  | Old variable                                   | New variable                            |
+  | ---------------------------------------------- | --------------------------------------- |
+  | `--hop-richicon-placeholder-surface-color`     | `--hop-Richicon-placeholder-shadow`     |
+  | `--hop-richicon-placeholder-icon-color`        | `--hop-Richicon-placeholder-background` |
+  | `--hop-richicon-placeholder-icon-strong-color` | `--hop-Richicon-placeholder-fill`       |
+
+- Updated dependencies [c5c6b30]
+  - @hopper-ui/styled-system@2.0.0
+
 ## 1.10.1
 
 ### Patch Changes